+++ Prism/Functions/Install-PrivateModule.ps1
@@ -17,6 +17,17 @@ function Install-PrivateModule
         Use-CallerPreference -Cmdlet $PSCmdlet -SessionState $ExecutionContext.SessionState
 
         $pkgMgmtPrefs = Get-PackageManagementPreference
+
+        $repoByLocation = @{}
+        foreach ($repo in (Get-PSRepository))
+        {
+            $repoUrl = $repo.SourceLocation
+            $repoByLocation[$repoUrl] = $repo.Name
+
+            # Ignore slashes at the end of URLs.
+            $trimmedRepoUrl = $repoUrl.TrimEnd('/')
+            $repoByLocation[$trimmedRepoUrl] = $repo.Name
+        }
     }
 
     process
@@ -26,18 +37,8 @@ function Install-PrivateModule
             $Configuration | Update-ModuleLock | Format-Table
         }
 
-        $repoByLocation = @{}
-        Get-PSRepository |
-            ForEach-Object {
-                $repoUrl = $_.SourceLocation
-                $repoByLocation[$repoUrl] = $_.Name
-
-                # Ignore slashes at the end of URLs.
-                $trimmedRepoUrl = $repoUrl.TrimEnd('/')
-                $repoByLocation[$trimmedRepoUrl] = $_.Name
-            }
-
-        $privateModulePathWildcard = Join-Path -Path $config.PSModulesPath -ChildPath '*'
+        $installDirPath = $Configuration.InstallDirectoryPath
+        $privateModulePathWildcard = Join-Path -Path $installDirPath -ChildPath '*'
         $locks = Get-Content -Path $Configuration.LockPath | ConvertFrom-Json
         $locks | Add-Member -Name 'PSModules' -MemberType NoteProperty -Value @() -ErrorAction Ignore
 
@@ -60,10 +61,7 @@ function Install-PrivateModule
                     return "$($this.Version)$($prerelease)"
                 }
 
-            $savePath =
-                Join-Path -Path $Configuration.File.DirectoryName -ChildPath $Configuration.PSModulesDirectoryName
-
-            $installedModule = $installedModules | Where-Object SemVer -EQ $module.version 
+            $installedModule = $installedModules | Where-Object 'SemVer' -EQ $module.version
             if (-not $installedModule)
             {
                 $sourceUrl = $module.repositorySourceLocation
@@ -83,20 +81,39 @@ function Install-PrivateModule
                     continue
                 }
 
-                if (-not (Test-Path -Path $savePath))
+                if (-not (Test-Path -Path $installDirPath))
                 {
-                    New-Item -Path $savePath -ItemType 'Directory' -Force | Out-Null
+                    New-Item -Path $installDirPath -ItemType 'Directory' -Force | Out-Null
                 }
 
                 Save-Module -Name $module.name `
-                            -Path $savePath `
+                            -Path $installDirPath `
                             -RequiredVersion $module.version `
                             -AllowPrerelease `
                             -Repository $repoName `
                             @pkgMgmtPrefs
+
+                # How many versions of this module will we be installing?
+                $moduleVersionCount = ($locks.PSModules | Where-Object 'Name' -EQ $module.name | Measure-Object).Count
+
+                # PowerShell has a 10 directory limit for nested modules, so reduce the number of nested directories
+                # when installing a nested module by installing directly in the module root directory and moving
+                # everything out of the version module directory.
+                if ($Configuration.Nested -and $moduleVersionCount -eq 1)
+                {
+                    $modulePath = Join-Path -Path $installDirPath -ChildPath $module.name
+                    $versionDirName = $module.version
+                    if ($versionDirName -match '^(\d+\.\d+\.\d+)')
+                    {
+                        $versionDirName = $Matches[1]
+                    }
+                    $moduleVersionPath = Join-Path -Path $modulePath -ChildPath $versionDirName
+                    Get-ChildItem -Path $moduleVersionPath -Force | Move-Item -Destination $modulePath
+                    Get-Item -Path $moduleVersionPath | Remove-Item
+                }
             }
 
-            $modulePath = Join-Path -Path $savePath -ChildPath $module.name | Resolve-Path -Relative
+            $modulePath = Join-Path -Path $installDirPath -ChildPath $module.name | Resolve-Path -Relative
             $installedModule = [pscustomobject]@{
                 Name = $module.name;
                 Version = $module.version;