+++ Tests/install.Tests.ps1
@@ -26,6 +26,32 @@ BeforeAll {
 }
 "@
 
+    function GivenFile
+    {
+        param(
+            [Parameter(Mandatory)]
+            [String] $Named,
+
+            [String] $WithContent,
+
+            [String] $In = $script:testRoot
+        )
+
+        $filePath = Join-Path -Path $In -ChildPath $Named
+        $dirPath = $filePath | Split-Path -Parent
+        if (-not (Test-Path -Path $dirPath))
+        {
+            New-Item -Path $dirPath -ItemType Directory
+        }
+
+        New-Item -Path $filePath -ItemType File
+
+        if ($WithContent)
+        {
+            $WithContent | Set-Content -Path $filePath -NoNewline
+        }
+    }
+
     function GivenPrismFile
     {
         param(
@@ -35,14 +61,7 @@ BeforeAll {
             [String] $In = $script:testRoot
         )
 
-        $path = 'prism.json'
-        if ($In)
-        {
-            New-Item -Path $In -ItemType 'Directory' -Force | Out-Null
-            $path = Join-Path -Path $In -ChildPath $path
-        }
-
-        $Contents | Set-Content -Path $path -NoNewline
+        GivenFile 'prism.json' -In $In -WithContent $Contents
     }
 
     function GivenLockFile
@@ -54,31 +73,41 @@ BeforeAll {
             [String] $In = $script:testRoot
         )
 
-        $path = 'prism.lock.json'
-        if ($In)
-        {
-            New-Item -Path $In -ItemType 'Directory' -Force | Out-Null
-            $path = Join-Path -Path $In -ChildPath $path
-        }
-
-        $Contents | Set-Content -Path $path -NoNewline
+        GivenFile 'prism.lock.json' -In $In -WithContent $Contents
     }
 
     function ThenInstalled
     {
+        [CmdletBinding(DefaultParameterSetName='NonNested')]
         param(
-            [Parameter(Mandatory)]
+            [Parameter(Mandatory, Position=0)]
             [hashtable] $Module,
 
-            [String] $In = $script:testRoot,
+            [String] $In,
 
-            [String] $UsingDirName = 'PSModules'
+            [Parameter(ParameterSetName='NonNested')]
+            [String] $UsingDirName,
+
+            [Parameter(Mandatory, ParameterSetName='Nested')]
+            [switch] $AsNestedModule
         )
 
-        $savePath = $UsingDirName
-        if ($In)
+        if (-not $In)
         {
-            $savePath = Join-Path -Path $In -ChildPath $savePath
+            $In = $script:testRoot
+        }
+
+        if (-not $UsingDirName)
+        {
+            $UsingDirName = 'PSModules'
+        }
+
+        $isNestedModule = $PSCmdlet.ParameterSetName -eq 'Nested'
+
+        $savePath = $In
+        if (-not $isNestedModule)
+        {
+            $savePath = Join-Path -Path $In -ChildPath $UsingDirName
         }
 
         # Make sure *only* the modules we requested are installed.
@@ -91,6 +120,11 @@ BeforeAll {
                 $expectedCount += 1
                 $version,$prerelease = $semver -split '-'
                 $manifestPath = Join-Path -Path $modulePath -ChildPath $version
+                if ($isNestedModule -and ($Module[$moduleName] | Measure-Object).Count -eq 1)
+                {
+                    $manifestPath | Should -Not -Exist -Because 'should remove version directory for nested module'
+                    $manifestPath = $manifestPath | Split-Path -Parent
+                }
                 $manifestPath = Join-Path -Path $manifestPath -ChildPath "$($moduleName).psd1"
                 $manifestPath | Should -Exist
                 $manifest =
@@ -109,7 +143,13 @@ BeforeAll {
             }
         }
 
-        Get-ChildItem -Path "$($savePath)\*\*\*.psd1" -ErrorAction Ignore |
+        $path = "${savePath}\*\*\*.psd1"
+        if ($isNestedModule -and ($Module[$moduleName] | Measure-Object).Count -eq 1)
+        {
+            $path = "${savePath}\*\*.psd1"
+        }
+
+        Get-ChildItem -Path $path -ErrorAction Ignore |
             Select-Object -ExpandProperty 'DirectoryName' |
             Select-Object -Unique |
             Should -HaveCount $expectedCount
@@ -538,5 +578,73 @@ Describe 'prism install' {
 "@
         WhenInstalling -WithParameters @{ Name = 'Carbon'}
         ThenNotInstalled 'Carbon'
+    }
+
+    Context 'installing nested module' {
+        It 'reduces nesting in module with <_> file' -ForEach @('module.psd1', 'module.psm1') {
+            GivenFile $_
+            GivenPrismFile '{}'
+            GivenLockFile @'
+{
+    "PSModules": [
+        {
+            "name": "NoOp",
+            "version": "1.0.0",
+            "repositorySourceLocation": "https://www.powershellgallery.com/api/v2/"
+        }
+    ]
+}
+'@
+            WhenInstalling
+            ThenInstalled @{ 'NoOp' = '1.0.0' } -AsNestedModule
+        }
+
+        It 'does not reinstall if already installed' {
+            GivenFile 'module.psd1'
+            GivenPrismFile @"
+{
+    "PSModules": [
+        {
+            "Name": "NoOp",
+            "Version": "1.0.0"
+        }
+    ]
+}
+"@
+            WhenInstalling
+            ThenInstalled @{ 'NoOp' = '1.0.0' } -AsNestedModule
+            Mock -CommandName 'Save-Module' -ModuleName 'Prism'
+            WhenInstalling
+            Should -Invoke 'Save-Module' -ModuleName 'Prism' -Times 0 -Exactly
+        }
+
+        It 'installs multiple versions' {
+            GivenFile 'module.psd1'
+            GivenPrismFile '{}' # install only cares about prism.lock.json
+            GivenLockFile @"
+{
+    "PSModules": [
+        { "name": "Carbon", "version": "2.11.1", "repositorySourceLocation": "$($script:defaultLocation)" },
+        { "name": "Carbon", "version": "2.11.0", "repositorySourceLocation": "$($script:defaultLocation)" }
+    ]
+}
+"@
+            WhenInstalling
+            ThenInstalled @{ 'Carbon' = @('2.11.1', '2.11.0') } -AsNestedModule
+        }
+
+        It 'installs prerelease modules' {
+            GivenFile 'module.psd1'
+            GivenPrismFile '{}'
+            GivenLockFile @"
+{
+    "PSModules": [
+        { "name": "NoOp", "version": "1.0.0-alpha26", "repositorySourceLocation": "$($script:defaultLocation)" }
+    ]
+}
+"@
+            WhenInstalling
+            ThenInstalled @{ 'NoOp' = '1.0.0-alpha26' } -AsNestedModule
+        }
     }
 }